// src/types/ICrypto.ts
export type CryptoResponse = {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  fully_diluted_valuation: number;
  total_volume: number;
  high_24h: number;
  low_24h: number;
  price_change_24h: number;
  price_change_percentage_24h: number;
  market_cap_change_24h: number;
  market_cap_change_percentage_24h: number;
  circulating_supply: number;
  total_supply: number;
  max_supply: number;
  ath: number;
  ath_change_percentage: number;
  ath_date: Date;
  atl: number;
  atl_change_percentage: number;
  atl_date: Date;
  roi?: any;
  last_updated: Date;
  price_change_percentage_1h_in_currency: number;
};

export type CryptoSearchResponse = {
  id: string;
  name: string;
  symbol: string;
  market_cap: number;
  thumb: string;
  large: string;
};

export type CryptoChartResponse = {
  [key: string]: number;
};

export type CryptoTrendingResponse = {
  id: string;
  coin_id: number;
  name: string;
  symbol: string;
  market_cap_rank: number;
  thumb: string;
  small: string;
  large: string;
  slug: string;
  price_btc: number;
  score: number;
};

export type CryptoTrendingItemResponse = {
  item: CryptoTrendingResponse;
};